scripts/qemu.py: allow to launch the VM without a monitor
QEMUMachine launches the VM with a monitor enabled, afterwards
a qmp connection is attempted on _post_launch(). In case
the QEMU process exits with an error, qmp.accept() reaches
timeout and raises an exception.

But sometimes you don't need that monitor. As an example,
when a test launches the VM expects its immediate crash,
and only intend to check the process's return code. In this
case the fact that launch() tries to establish the qmp
connection (ending up in an exception) is troublesome.

So this patch adds the set_qmp_monitor() that allow to
launch the VM without creating the monitor machinery. The
method can be used to re-enable the monitor eventually.

Tested this change with the following Avocado test:

from avocado_qemu import Test

class DisableQmp(Test):
    """
    :avocado: enable
    """
    def test(self):
        self.vm.add_args('--foobar')
        self.vm.set_qmp_monitor(disabled=True)
        self.vm.launch()
        self.vm.wait()
        self.assertEqual(self.vm.exitcode(), 1)
        self.vm.shutdown()

        self.vm.set_qmp_monitor(disabled=False)
        self.vm._args.remove('--foobar') # Hack
        self.vm.launch()
        res = self.vm.command('human-monitor-command',
                              command_line='info version')
        self.assertRegexpMatches(res, r'^(\d+\.\d+\.\d)')

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>