scripts/qemu.py: allow to launch QEMUMachine without qmp
Currently QEMUMachine is launched with monitor enabled and
a qmp connection is established on _post_launch(), unless
the process exits with error and thus qmp.accept() reaches
timeout.

However it does not always need that qmp connection nor
wait for an expected timeout. As an example, when a test
launches the VM process expecting its immediate crash,
and you only want to check the return code.

So this change adds an argument to launch() that allows to
launch the VM without creating the entire qmp machinery.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>