scripts/qemu.py: allow to launch the VM without a monitor
QEMUMachine launches the VM with a monitor enabled, afterwards
a qmp connection is attempted on _post_launch(). In case
the QEMU process exits with an error, qmp.accept() reaches
timeout and raises an exception.

But sometimes you don't need that monitor. As an example,
when a test launches the VM expects its immediate crash,
and only intend to check the process's return code. In this
case the fact that launch() tries to establish the qmp
connection (ending up in an exception) is troublesome.

So this patch adds the disable_qmp() that allow to
launch the VM without creating the monitor machinery.

Tested this change with the following Avocado test:

from avocado_qemu import Test
class DisableQmp(Test):
    """
    :avocado: enable
    """
    def test(self):
        self.vm.add_args('--foobar')
        self.vm.disable_qmp()
        self.vm.launch()
        self.vm.wait()
        self.assertEqual(self.vm.exitcode(), 1)

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>