hw/core: add qom getter for kernel-irqchip property
Allows to access the kernel-irqchip property of a Machine
Class object via QOM get method.

Before this patch the property cannot be read although it is
listed by qom-list:

qemu-system-x86_64 -M q35,accel=kvm,kernel-irqchip=split (...)
-> {"execute": "qom-list", "arguments": {"path": "/machine"}}
<- {"return": [{"name": "type", "type": "string"}, (...) , {"name": "kernel-irqchip", "type": "on|off|split"} (...)}
-> {"execute": "qom-get", "arguments": {"path": "/machine", "property": "kernel-irqchip"}}
<- {"error": {"class": "GenericError", "desc": "Insufficient permission to perform this operation"}}

It is implemented the machine_get_kernel_irqchip() method,
which evaluates the kernel_irqchip_allowed, *_required, and
*_split fields to determine the value of kernel-irqchip. Note: we
assume there is not inconsistency on the value of those fields.

Then with this change in place, it works as expected:

qemu-system-x86_64 -M q35,accel=kvm,kernel-irqchip=split (...)
-> {"execute": "qom-get", "arguments": {"path": "/machine", "property": "kernel-irqchip"}}
<- {"return": "split"}

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>