Acceptance test: add the "accel" tag
Some test cases may boot a VM with accelerator that isn't actually
enabled on the QEMU binary or present in the host. In this case
the test case is gonna fail miserably, unless it can be skipped.

This change introduces the "accel" tag, used to mark the test
case requires a given accelerator(s). It was implemented a
mechanism to check the given accelerator is available, and if not
then the test case is skipped.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>