tests/acceptance: Introduce the BootKernelTest class
Some acceptance tests rely on booting a Linux Kernel and then
finding messages in the console to check if it was loaded (or not)
correctly. There is duplicated code among those tests though. So
this change creates a BootKernelTest class that provides the common
logic for tests that require Kernel boot.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>