tests/acceptance: remove self.arch from avocado_qemu.Test
The self.arch of avocado_qemu.Test object is obtained
from either test tags or via Avocado parameter, and both can be not
set which results in self.arch equals to None. In theory its value
could be guessed from the target binary filename, but may not be
possible  if qemu_bin parameter is passed. Therefore, self.arch is
not guaranted to match the target binary or can be None. Since its
use is then misleading, it is error prone to expose self.arch to
tests usage.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>