tests/acceptance: Handle machine type for ARM target
Some tests are meant arch-independent and as such they don't set
the machine type (i.e. relying to defaults) on launched VMs. The arm
targets, however, don't provide any default machine so tests fail.

This patch adds a logic on the base Test class so that machine type
is set to 'virt' when:
   a) The test case doesn't have arch:aarch64 or arch:arm tag. Here
      I assume that if the test was tagged for a specific arch then
      the writer took care of setting a machine type.
   b) The target binary arch is any of aarch64 or arm. Note:
      self.target_arch can end up None if qemu_bin is passed by
      Avocado parameter and the filename doesn't match expected
      format. In this case the test will fail.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>