travis.yml: avocado: Show log of not passed tests only
The acceptance tests build on Travis is configured to print
the entire Avocado's job log in case any test fail. Usually one is
interested on failed tests only though. So this change the
configuration in order to show the log of tests which status is
different from 'PASS' and 'SKIP'.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>