travis.yml: Detach build and test steps
Currently build and test commands are a single step in a
Travis's `script` block. Thus, in order to see the output
of the tests one needs to scroll down the log to find where
the build messages ended. If they were in different steps
then Travis would put their outputs each on toggled
text boxes. So this change is made to detach those commands
to ease the visualization of the output.

Note that all steps on the `script` block is executed regardless
if one previous has failed. To overcome it, let's save the
return code of the build then check whether succeed or failed on
the test step.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>